fix files i broke with my silly crypto rename
