+++ src/Lyn.Protocol.Tests/Bolt2/AcceptChannelMessageServiceTests.cs
@@ -9,7 +9,7 @@ using Lyn.Protocol.Bolt3;
 using Lyn.Protocol.Bolt9;
 using Lyn.Protocol.Common;
 using Lyn.Protocol.Common.Blockchain;
-using Lyn.Protocol.Common.Hashing;
+using Lyn.Protocol.Common.Crypto;
 using Lyn.Protocol.Common.Messages;
 using Lyn.Protocol.Connection;
 using Lyn.Types;
